Refactor external review miss derivation
